% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_test.R
\name{print.mvn_test}
\alias{print.mvn_test}
\title{Print Method for \code{mvn_test} Objects}
\usage{
\method{print}{mvn_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mvn_test}.}

\item{...}{Further arguments (currently ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print the result of the test.
}
