% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canodist.R
\name{canodist}
\alias{canodist}
\title{Canonical distribution}
\usage{
canodist(part, nsample, pvariates, iterations)
}
\arguments{
\item{part}{.}

\item{nsample}{.}

\item{pvariates}{.}

\item{iterations}{.}
}
\description{
This function calculates
}
\examples{
canodist(part = 2, nsample = 100, pvariates = 4, iterations = 2)

}
\references{
ref
}
