% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Canodist.R
\name{canodist}
\alias{canodist}
\title{Canonical Empirical Distribution}
\usage{
canodist(part, nsample, pvariates, iterations)
}
\arguments{
\item{part}{Number of partitions.}

\item{nsample}{Sample size.}

\item{pvariates}{Number of variables.}

\item{iterations}{Number of iterations for simulating values from the distribution and finding the quantiles. Default is \code{10000}.}
}
\description{
This function calculates the empirical distribution of the pivotal random
variable that can be used to perform inferential procedures and
test for the regression of one set of variables on the other based on the
released Single Synthetic data generated under Plug-in Sampling, assuming
that the original dataset is normally distributed. We define
\deqn{T_4^\star|\boldsymbol{\Delta} =
\frac{(|\boldsymbol{S}^{\star}_{12}
(\boldsymbol{S}^{\star}_{22})^{-1}-\boldsymbol{\Delta})
\boldsymbol{S}^{\star}_{22}(\boldsymbol{S}^{\star}_{12})
(\boldsymbol{S}^{\star}_{22})^{-1}-\boldsymbol{\Delta})^\top|}
{|\boldsymbol{S}^{\star}_{11.2}|}}
where \eqn{\boldsymbol{S}^\star = \sum_{i=1}^n (v_i - \bar{v})(v_i - \bar{v})^{\top}},
\eqn{v_i} is the \eqn{i}th observation of the synthetic dataset,
considering \eqn{\boldsymbol{S}^\star} partitioned as
\deqn{\boldsymbol{S}^{\star}=\left[\begin{array}{lll}
\boldsymbol{S}^{\star}_{11}& \boldsymbol{S}^{\star}_{12}\\
\boldsymbol{S}^{\star}_{21} & \boldsymbol{S}^{\star}_{22}
\end{array}\right].}
For \eqn{\Delta = \boldsymbol{\Sigma}_{12}\boldsymbol{\Sigma}_{22}^{-1}},
its distribution is stochastic equivalent to
\deqn{\frac{|\boldsymbol{\Omega}_{12}\boldsymbol{\Omega}_{22}^{-1}
\boldsymbol{\Omega}_{21}|}{|\boldsymbol{\Omega}_{11}-\boldsymbol{\Omega}_{12}
\boldsymbol{\Omega}_{22}^{-1}\boldsymbol{\Omega}_{21}|}}
where \eqn{\boldsymbol{\Omega}} are all independent Wishart random variables,
\eqn{\boldsymbol{\Omega} \sim \mathcal{W}_p(n-1, \frac{\boldsymbol{W}}{n-1})}
where
\eqn{\boldsymbol{W} \sim \mathcal{W}_p(n-1, I_p)} and
\eqn{\boldsymbol{\Omega}} partitioned in the same way as
\eqn{\boldsymbol{S}^{\star}}.
To test \eqn{\mathcal{H}_0: \boldsymbol{\Delta} =\Delta_0}, compute the value
 of \eqn{T_{4}^\star}, \eqn{\widetilde{T_{4}^\star}},  with the observed
 values and reject the null hypothesis if
\eqn{\widetilde{T_{4}^\star}>t^\star_{4,1-\alpha}} for
\eqn{\alpha}-significance level, where \eqn{t^\star_{4,\gamma}} is the
\eqn{\gamma}th percentile of \eqn{T_4^\star}.
}
\examples{
library(MASS)
n_sample = 100
p = 4
mu <- c(1,2,3,4)
Sigma = matrix(c(1,0.5,0,0,
                  0.5,2,0,0,
                  0,0,3,0.2,
                  0, 0, 0.2,4), nr = 4, nc = 4, byrow = TRUE)
df = mvrnorm(n_sample, mu = mu, Sigma = Sigma)
df_s = simSynthData(df)

part = 2
Sigma_12 = partition(Sigma4,nrows = part, ncol = part)[[2]]
Sigma_22 = partition(Sigma4,nrows = part, ncol = part)[[4]]
Delta0 = Sigma_12 \%*\% solve(Sigma_22)

Sstar = cov(df_s)
Sstar_11 = partition(Sstar,nrows = part, ncol = part)[[1]]
Sstar_12 = partition(Sstar,nrows = part, ncol = part)[[2]]
Sstar_21 = partition(Sstar,nrows = part, ncol = part)[[3]]
Sstar_22 = partition(Sstar,nrows = part, ncol = part)[[4]]


DeltaEst = Sstar_12 \%*\% solve(Sstar_22)
Sstar11_2 = Sstar_11 - Sstar_12 \%*\% solve(Sstar_22) \%*\% Sstar_21


T4_star = det((DeltaEst-Delta0)\%*\%Sstar_22\%*\%t(DeltaEst-Delta0))/det(Sstar11_2)

alpha = 0.05
T4 <- canodist(part = part, nsample = n, pvariates = p, iterations = 1000)
q975 <- quantile(T4, 1-alpha)

T4_star > q975 #False means Delta != Delta0
}
\references{
Klein, M., Moura, R. and Sinha, B. (2021). Multivariate Normal Inference based on Singly Imputed Synthetic Data under Plug-in Sampling. Sankhya B 83, 273–287.
}
