% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inddist.R
\name{Inddist}
\alias{Inddist}
\title{Independence Empirical Distribution}
\usage{
Inddist(part, nsample, pvariates, iterations)
}
\arguments{
\item{part}{.}

\item{nsample}{.}

\item{pvariates}{.}

\item{iterations}{.}
}
\description{
This function calculates
}
\examples{
Inddist(part = 2, nsample = 100, pvariates = 4, iterations = 2)

}
\references{
Klein, M., Moura, R. and Sinha, B. (2021). Multivariate Normal Inference based on Singly Imputed Synthetic Data under Plug-in Sampling. Sankhya B 83, 273–287.
}
