% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inddist.R
\name{Inddist}
\alias{Inddist}
\title{Inddist distribution}
\usage{
Inddist(part, nsample, pvariates, iterations)
}
\arguments{
\item{part}{.}

\item{nsample}{.}

\item{pvariates}{.}

\item{iterations}{.}
}
\description{
This function calculates
}
\examples{
Inddist(part = 2, nsample = 100, pvariates = 4, iterations = 2)

}
\references{
ref
}
