% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Split a matrix into blocks}
\usage{
partition(mat, nrows, ncols)
}
\arguments{
\item{nrows}{either a Positive integer indicating the  number of rows blocks.}

\item{ncols}{either a Positive integer indicating the  number of columns blocks.}

\item{Matrix}{a matrix to split .}
}
\description{
This function Split a matrix into a list of blocks (either by rows and  columns).
}
\examples{
df = matrix(c(1,0.5,0,0,
                  0.5,2,0,0,
                  0,0,3,0.2,
                  0, 0, 0.2,4), nr = 4, nc = 4, byrow = TRUE)
partition(df,2,2)
}
