% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Split a matrix into blocks}
\usage{
partition(Matrix, nrows, ncols)
}
\arguments{
\item{Matrix}{a matrix to split .}

\item{nrows}{positive integer indicating the number of rows blocks.}

\item{ncols}{positive integer indicating the number of columns blocks.}
}
\value{
a list of partitioned submatrices
}
\description{
This function split a matrix into a list of blocks (either by rows and columns).
}
\examples{
Mat = matrix(c(1,0.5,0,0,
                  0.5,2,0,0,
                  0,0,3,0.2,
                  0, 0, 0.2,4), nrow = 4, ncol = 4, byrow = TRUE)
partition(Matrix = Mat, nrows = 2, ncols = 2)
}
