% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSynthData.R
\name{simSynthData}
\alias{simSynthData}
\title{Plug-in Sampling Single Synthetic Dataset Generation}
\usage{
simSynthData(X, n_imp = dim(X)[1])
}
\arguments{
\item{X}{kjk}

\item{n_imp}{xx}
}
\description{
This function calculates
}
\examples{
library(MASS)
n_sample = 100
mu=c(0,0,0,0)
Sigma=diag(1,4,4)
# Create original simulated dataset
df_o = mvrnorm(n_sample, mu, Sigma)
# Create singly imputed synthetic dataset
df_s = simSynthData(df_o)
#Estimators synthetic
mean_s <- colMeans(df_s)
S_s <- (t(df_s)- mean_s) \%*\% t(t(df_s)- mean_s)
# careful about this computation
# mean_o is a column vector and if you are thinking as n X p matrices and
# row vectors you should be aware of this.
print(mean_s)
print(S_s/(dim(df_s)[1]-1))
}
\references{
ref
}
