% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sphdist.R
\name{Sphdist}
\alias{Sphdist}
\title{sphdist}
\usage{
Sphdist(nsample, pvariates, iterations)
}
\arguments{
\item{nsample}{.}

\item{pvariates}{.}

\item{iterations}{.}
}
\description{
This function calculates
}
\examples{
Sphdist(nsample = 100, pvariates = 4, iterations = 2)

}
\references{
ref
}
